@@ -4,6 +4,7 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
+using System.Net.Http;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading;
@@ -12,18 +13,27 @@ using System.Threading.Tasks;
 namespace WordProcessor
 {
     /// <summary>
-    /// Class to download content from given url traversing upto a certain level of depth of the hyperlinks of same domain
+    /// Class to download content from given url traversing up to a certain level of depth of the hyperlinks of same domain
     /// </summary>
-    public class Downloader
+    /// <remarks>
+    /// Constructor for Downloaded
+    /// </remarks>
+    /// <param name="depth"></param>
+    public partial class Downloader(byte depth)
     {
         /// <summary>
-        /// Constructor for Downloaded
+        /// Regex to find hyperlink inside html content
         /// </summary>
-        /// <param name="depth"></param>
-        public Downloader(byte depth)
-        {
-            maxDepth = depth;
-        }
+        static readonly Regex hyperlinkPattern = hyperLinkExp();
+
+        readonly byte maxDepth = depth;
+
+        private float presentLevel = 0;
+        /// <summary>
+        /// Present Level
+        /// </summary>
+        public float PresentLevel => presentLevel;
+
         /// <summary>
         /// Outputs string stripping off html tags and extra spaces
         /// </summary>
@@ -33,15 +43,11 @@ namespace WordProcessor
         {
             var doc = new HtmlDocument();
             doc.LoadHtml(input ?? "");
-            return Regex.Replace(doc.DocumentNode.InnerText.Trim(), @"\s+", " ");
+            return spaceExp().Replace(doc.DocumentNode.InnerText.Trim(), " ");
         }
-        //public static string CleanHtml(string s) => s.Replace( Regex.Replace(s, "<[^>]*>", string.Empty, RegexOptions.Multiline);
-
-        readonly byte maxDepth = 3;
-
         /// <summary>
-        /// Synshronous call to fetch html cleaned content from given url.
-        /// It traverses upto a certain level of depth of the hyperlinks of same domain
+        /// Synchronous call to fetch html cleaned content from given url.
+        /// It traverses up to a certain level of depth of the hyperlinks of same domain
         /// </summary>
         /// <param name="link"></param>
         /// <param name="visited">Keep track already visited links</param>
@@ -55,10 +61,10 @@ namespace WordProcessor
             if (visited.Contains(linkString))
                 return string.Empty;
             string html = string.Empty;
-            using var wc = new WebClient();
+            using var wc = new HttpClient();
             try
             {
-                html = wc.DownloadString(link);
+                html = wc.GetStringAsync(link).Result;
             }
             catch (WebException)
             {
@@ -73,7 +79,7 @@ namespace WordProcessor
             level++;
             var hyperLinks = GetHyperLinks(html, link.Host, visited);
             html = CleanHtml(html);
-            if (!hyperLinks.Any())
+            if (hyperLinks.Count == 0)
                 return html;
             var htmlBuilder = new StringBuilder(html);
             do
@@ -82,15 +88,20 @@ namespace WordProcessor
                 var txt = GetTextFromUrl(hl, visited, level);
                 if (!string.IsNullOrWhiteSpace(txt))
                     htmlBuilder.AppendLine(txt);
-            } while (hyperLinks.Any());
+            } while (hyperLinks.Count != 0);
             return htmlBuilder.ToString();
         }
-
-        public float PresentLevel => presentLevel;
-        private float presentLevel = 0;
+        /// <summary>
+        /// Synchronous call to fetch html cleaned content from given url.
+        /// </summary>
+        /// <param name="link"></param>
+        /// <returns></returns>
+        /// <seealso cref="GetTextFromUrl(Uri, List{string}, byte)"/>
+        /// <seealso cref="GetTextFromUrlAsync(Uri)"/>
+        public string GetTextFromUrl(Uri link) => GetTextFromUrl(link, []);
 
         /// <summary>
-        /// Asynshronous multithreaded call to fetch html cleaned content from given url.
+        /// Asynchronous multithreaded call to fetch html cleaned content from given url.
         /// </summary>
         /// <param name="link"></param>
         /// <param name="visited">Keep track already visited links and corresponding contents</param>
@@ -103,14 +114,14 @@ namespace WordProcessor
             var linkString = link.ToString().ToLower();
             if (visited.ContainsKey(linkString))
                 return;
-            using var wc = new WebClient();
+            using var wc = new HttpClient();
             string html = string.Empty;
             var cleaned = string.Empty;
             try
             {
-                html = await wc.DownloadStringTaskAsync(link);
+                html = await wc.GetStringAsync(link).ConfigureAwait(false);
             }
-            catch (WebException)
+            catch (Exception)
             {
                 return;
             }
@@ -126,7 +137,7 @@ namespace WordProcessor
             level++;
             var lastProgress = 1.0f * level / maxDepth;
             var hyperLinks = GetHyperLinks(html, link.Host, visited.Keys);
-            if (!hyperLinks.Any())
+            if (hyperLinks.Count == 0)
             {
                 if (presentLevel < lastProgress)
                     Interlocked.Exchange(ref presentLevel, lastProgress);
@@ -143,32 +154,8 @@ namespace WordProcessor
                 t.Wait();
             });
         }
-
         /// <summary>
-        /// Regex to find hyperlink inside html content
-        /// </summary>
-        static readonly Regex hyperlinkPattern = new Regex(@"(?<=\<a.*\bhref\=\"")[^\""]*", RegexOptions.IgnoreCase | RegexOptions.ExplicitCapture);
-
-        /// <summary>
-        /// Find hyperlinks in a page for the the given hostname excluding the already visited link
-        /// </summary>
-        /// <param name="html"></param>
-        /// <param name="host"></param>
-        /// <param name="visited">Collection of already visited and to be excluded links</param>
-        /// <returns></returns>
-        private static Queue<Uri> GetHyperLinks(string html, string host, IEnumerable<string> visited) => new Queue<Uri>(hyperlinkPattern.Matches(html).Select(p => p.Value.ToLower()).Where(p => p.Contains(host) && !visited.Contains(p)).Select(p => new Uri(p)));
-
-        /// <summary>
-        /// Synshronous call to fetch html cleaned content from given url.
-        /// </summary>
-        /// <param name="link"></param>
-        /// <returns></returns>
-        /// <seealso cref="GetTextFromUrl(Uri, List{string}, byte)"/>
-        /// <seealso cref="GetTextFromUrlAsync(Uri)"/>
-        public string GetTextFromUrl(Uri link) => GetTextFromUrl(link, new List<string>());
-
-        /// <summary>
-        /// Asynshronous multithreaded call to fetch html cleaned content from given url.
+        /// ASynchronous multithreaded call to fetch html cleaned content from given url.
         /// </summary>
         /// <param name="link"></param>
         /// <returns></returns>
@@ -181,5 +168,19 @@ namespace WordProcessor
             await GetTextFromUrlAsync(link, visited);
             return String.Join(" ", visited.Select(p => p.Value));
         }
+
+        /// <summary>
+        /// Find hyperlinks in a page for the the given hostname excluding the already visited link
+        /// </summary>
+        /// <param name="html"></param>
+        /// <param name="host"></param>
+        /// <param name="visited">Collection of already visited and to be excluded links</param>
+        /// <returns></returns>
+        private static Queue<Uri> GetHyperLinks(string html, string host, IEnumerable<string> visited) => new(hyperlinkPattern.Matches(html).Select(p => p.Value.ToLower()).Where(p => p.Contains(host) && !visited.Contains(p)).Select(p => new Uri(p)));
+        [GeneratedRegex(@"(?<=\<a.*\bhref\=\"")[^\""]*", RegexOptions.IgnoreCase | RegexOptions.ExplicitCapture, "en-IN")]
+        private static partial Regex hyperLinkExp();
+
+        [GeneratedRegex(@"\s+")]
+        private static partial Regex spaceExp();
     }
 }
